+++ src/typings/catalog.ts
@@ -240,3 +240,12 @@ export interface Category {
   LinkId: string
   HasChildren: boolean
 }
+
+export interface GetBrandResponse {
+  id: number
+  name: string
+  imageUrl: string
+  isActive: boolean
+  title: string
+  metaTagDescription: string
+}