+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `getBrandById` method to the Catalog client
+- `listInventoryBySku` method to the Logistics client
+
+### Changed
+- Added `from` and `to` query params to `getProductsAndSkus` in the Catalog client
 
 ## [2.14.0] - 2021-07-07
 ### Added